Move password policy to appropriate place
