+++ backend/login/login.py
@@ -1,0 +1,21 @@
+from django.contrib.auth import authenticate
+from rest_framework.authtoken.models import Token
+from rest_framework.response import Response
+from rest_framework.status import HTTP_200_OK, HTTP_400_BAD_REQUEST
+from rest_framework.views import APIView
+
+
+class LoginTokenAPIView(APIView):
+    authentication_classes = ()
+    permission_classes = ()
+
+    def post(self, request):
+        username = request.data.get("username")
+        password = request.data.get("password")
+        user = authenticate(username=username, password=password)
+        if not user:
+            return Response(
+                {'error': 'Invalid Credentials'}, status=HTTP_400_BAD_REQUEST
+            )
+        token, _ = Token.objects.get_or_create(user=user)
+        return Response({'token': token.key}, status=HTTP_200_OK)
