+++ backend/sensehel/urls.py
@@ -16,8 +16,7 @@ Including another URLconf
 from django.conf.urls import url
 from django.contrib import admin
 from django.urls import include
-
-from api.login import LoginTokenAPIView
+from login.login import LoginTokenAPIView
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls, name='admin'),
